+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Fixed not being able to clear exporter row in interface
 - Fixed interface having weird extraction / insertion rules. It can now take items from any side, or extract items from any side depending on the block that is inserting or extracting the item (like conduits).
 
+**Features**
+- Added config option to set the base energy usage of the Controller (default is 0)
+
 ### 0.8.12
 **Bugfixes**
 - Fixed dupe bug when shift clicking output slot in grid